airflow/requirements.txt
boto3==1.24.7
botocore==1.27.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
configparser==5.2.0
idna==3.3
jmespath==1.0.0
numpy==1.22.4
pandas==1.4.2
praw==7.7.1
prawcore==2.3.0
psycopg2-binary==2.9.3
pyarrow==8.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
pendulum